:rotating_light: MASSIVE: Added original "Xane.com" website!
For anyone that wants to see it for themselves, here it is, the original Xane.com, found on my oldest laptop. Most pages weren't given the nasty, intended color scheme (common.css), but that'll come eventually. There's a lot of HTML files to update.

(Unfortunately, due to how the internet is moving on from it, the Flash content will be broken in most web browsers. A couple animations were converted into GIFs, but not all.