Replaced main page's header graphic.
The X Site has been...neglected for several years now, and I should do something about that, but it's honestly a useless website, and pretty much always has been. The only page that I visit regularly is the 'links" page, and it's the only one I recommend others to visit.

Despite this, I've replaced the main page's header image with a procedurally-generated 256 color image, made using Longbow Games' Texturizer.

Since this is a repository, I've included its project file in the "texturizerProjects" directory. You can open and modify it in that mid-90's program if you like. This repository doesn't have a license, and I don't know if any would be appropriate for a website's source code, but you may modify and use it as a base for your own images if you want.