Added section descriptions and revised list for accuracy.
The X Site is basically a corpse at this point; The website hasn't been updated in a while andwasn't made in an efficient, easy to update way (so every single page has its own redundant header and copy-pasted layout), so it's not only somewhat inaccurate, but it also includes broken links and it takes too much effort to give the illusion that all pages are using the same "code" or framework (when they're actually basic, self-contained HTML pages.

I'm slightly updating the list and the footer's copyright notice (when necessary), but this website has been another case of me not thinking of whether I SHOULD have done something, only thinking I COULD and immediately throwing something "cool" together like this website. The Google Sites version already exists and Google keeps a consistent style across all of the pages, but that didn't stop me from designing this weird website. I'm starting to think something needs to be done here so I can feel encouraged to maintain the website instead of seeing the HTML code and remembering that all of those tags have to be very similar on each page or the layout or search results may break.